Reduce log printing in IFW controller

+++ core/data/src/main/java/com/merxury/blocker/core/data/respository/component/LocalComponentRepository.kt
@@ -57,6 +57,7 @@ class LocalComponentRepository @Inject constructor(
             .map { it?.toComponentInfo() }
             .flowOn(ioDispatcher)
     }
+
     override fun getComponentList(packageName: String): Flow<List<ComponentInfo>> =
         appComponentDao.getByPackageName(packageName)
             .map { list ->
@@ -83,8 +84,10 @@ class LocalComponentRepository @Inject constructor(
             val diff = (latestComponents + cachedComponents).groupBy { it.componentName }
                 .filter { it.value.size == 1 }
                 .flatMap { it.value }
-            Timber.d("Found ${diff.size} components to delete for $packageName in type $type")
-            appComponentDao.delete(diff)
+            if (diff.isNotEmpty()) {
+                Timber.d("Found ${diff.size} components to delete for $packageName in type $type")
+                appComponentDao.delete(diff)
+            }
             appComponentDao.upsertComponentList(latestComponents)
             emit(Success(Unit))
         }
@@ -99,8 +102,10 @@ class LocalComponentRepository @Inject constructor(
                 val diff = (latest + cached).groupBy { it.componentName }
                     .filter { it.value.size == 1 }
                     .flatMap { it.value }
-                Timber.d("Found ${diff.size} components to delete for $packageName")
-                appComponentDao.delete(diff)
+                if (diff.isNotEmpty()) {
+                    Timber.d("Found ${diff.size} components to delete for $packageName")
+                    appComponentDao.delete(diff)
+                }
                 Timber.d("Update component list for $packageName, size: ${latest.size}")
                 appComponentDao.upsertComponentList(latest)
                 Success(Unit)
